refactor: clean up mappers using shared function
